pyproject.toml
[project]
name = "ir-datasets-webis"
authors = [
    {name = "Jan Heinrich Reimer", email = "heinrich.reimer@uni-jena.de"},
]
description = "Extension for accessing Webis datasets via ir_datasets."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "annotated-types~=0.6.0",
    "importlib-metadata~=7.0",
    "ir-datasets~=0.5.6",
    "pydantic~=2.7",
    "pyyaml~=6.0",
    "typing-extensions~=4.11",
]
dynamic = ["version"]

[project.optional-dependencies]
tests =[
    "bandit[toml]~=1.7",
    "mypy~=1.5",
    "pytest~=8.0",
    "pytest-cov~=5.0",
    "pytest-subtests~=0.12.1",
    "ruff~=0.3.0",
    "types-pyyaml~=6.0",
]

[project.urls]
"Homepage" = "https://github.com/heinrichreimer/ir-datasets-webis"
"Bug Tracker" = "https://github.com/heinrichreimer/ir-datasets-webis/issues"

[project.scripts]
ir_datasets_webis = "ir_datasets_webis:main_cli"
ir-datasets-webis = "ir_datasets_webis:main_cli"
irds-webis = "ir_datasets_webis:main_cli"

[build-system]
requires = [
    "setuptools~=69.0",
    "setuptools-scm~=8.0",    
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
exclude = [
    "build",
    "venv",
]

[tool.setuptools.package-data]
mypkg = ["*.txt", "*.md", "*.rst"]

[tool.setuptools_scm]

[tool.ruff]
exclude = [
    "build",
    "venv",
]

[tool.mypy]
exclude = [
    "build",
    "venv",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = [
    "build",
    "venv",
]

[tool.bandit.assert_used]
skips = ["**/test_*.py", "**/conftest.py"]

[tool.pytest.ini_options]
log_cli = "True"
